LayoutTests/imported/w3c/web-platform-tests/css/css-grid/grid-items/grid-minimum-size-grid-items-022-expected.txt


PASS .grid 1 
PASS .grid 2 
PASS .grid 3 
PASS .grid 4 
PASS .grid 5 
PASS .grid 6 
PASS .grid 7 
PASS .grid 8 
PASS .grid 9 
PASS .grid 10 
FAIL .grid 11 assert_equals: 
<div class="grid" style="grid-template-columns: minmax(auto, 25px);">
  <div data-expected-width="25">XXXXXXXXXX</div>
  <div data-expected-width="25"></div>
</div>
width expected 25 but got 100
FAIL .grid 12 assert_equals: 
<div class="grid" style="grid-template-columns: minmax(auto, 0px);">
  <div data-expected-width="0">XXXXXXXXXX</div>
  <div data-expected-width="0"></div>
</div>
width expected 0 but got 100
PASS .grid 13 
PASS .grid 14 
PASS .grid 15 
grid-template-columns: auto;
XXXXXXXXXX
grid-template-columns: 0px;
XXXXXXXXXX
grid-template-columns: 25px;
XXXXXXXXXX
grid-template-columns: minmax(auto, 0px);
XXXXXXXXXX
grid-template-columns: minmax(auto, 25px);
XXXXXXXXXX
grid-template-columns: minmax(auto, 0px); item width: 10px;
XXXXXXXXXX
grid-template-columns: minmax(auto, 25px); item width: 10px;
XXXXXXXXXX
grid-template-columns: minmax(auto, 0px); item margin width: 10px;
grid-template-columns: minmax(auto, 25px); item margin width: 10px;
grid-template-columns: minmax(auto, 0px); item padding width: 10px;
grid-template-columns: minmax(auto, 25px); item padding width: 10px;
grid-template-columns: minmax(auto, 0px); item border width: 10px;
grid-template-columns: minmax(auto, 25px); item border width: 10px;
grid-template-columns: minmax(auto, 0px); item width + margin + border + padding: 10px;
grid-template-columns: minmax(auto, 25px); item width + margin + border + padding: 10px;
